+++ praw/models/__init__.py
@@ -2,12 +2,13 @@
 
 from praw.models.auth import Auth
 from praw.models.front import Front
-from praw.models.helpers import DraftHelper, LiveHelper, MultiredditHelper, SubredditHelper
+from praw.models.helpers import AnnouncementHelper, DraftHelper, LiveHelper, MultiredditHelper, SubredditHelper
 from praw.models.inbox import Inbox
 from praw.models.list.draft import DraftList
 from praw.models.list.moderated import ModeratedList
 from praw.models.list.redditor import RedditorList
 from praw.models.list.trophy import TrophyList
+from praw.models.listing.announcement import AnnouncementListing
 from praw.models.listing.domain import DomainListing
 from praw.models.listing.generator import ListingGenerator
 from praw.models.listing.listing import Listing, ModeratorListing, ModmailConversationsListing
@@ -16,6 +17,7 @@ from praw.models.mod_action import ModAction
 from praw.models.mod_note import ModNote
 from praw.models.mod_notes import RedditModNotes, RedditorModNotes, SubredditModNotes
 from praw.models.preferences import Preferences
+from praw.models.reddit.announcement import Announcement
 from praw.models.reddit.collections import Collection
 from praw.models.reddit.comment import Comment
 from praw.models.reddit.draft import Draft